+++ console/migrations/m180105_021612_member_own_card.php
@@ -2,9 +2,6 @@
 
 use yii\db\Migration;
 
-/**
- * Class m180105_021612_member_own_card
- */
 class m180105_021612_member_own_card extends Migration
 {
     private $_table = '{{%member_own_card}}';
@@ -19,8 +16,8 @@ class m180105_021612_member_own_card extends Migration
         $this->createTable($this->_table, [
             'id'             => $this->primaryKey(),
             'user_id'        => $this->integer()->notNull(),
-            'original_money' => $this->integer()->notNull(),
-            'remain_money'   => $this->integer()->notNull(),
+            'original_money' => $this->double()->notNull(),
+            'remain_money'   => $this->double()->notNull(),
             'discount'       => $this->integer()->notNull(),
             'expire_at'      => $this->integer(),
             'created_at'     => $this->integer(),
